+++ ejercicios/avanzados/19_get_max_number.py
@@ -1,0 +1,8 @@
+import utils
+# from utils import get_max # Importa solo la funcion get_max
+
+numbers = [1, 2, 3, 4, 5]
+max_number = utils.get_max(numbers)
+
+
+print(f"El valor maximo es: {max_number}")