+++ lecciones/programacion-avanzada/25_packages.py
@@ -1,0 +1,4 @@
+# Los paquetes en python son colecciones de modulos que se encuentran en un mismo directorio
+# Los paquetes pueden contener modulos, subpaquetes y otros elementos
+# Los paquetes pueden ser importados y utilizados en otros archivos
+# Los paquetes pueden ser reutilizados en otros archivos